+++ mercury/mercury-library/models/src/main/scala/io/iohk/atala/mercury/model/error/package.scala
@@ -18,8 +18,11 @@ package object error {
 
   trait TransportError extends Exception with MercuryError
 
-  case class SendMessageError(cause: Throwable)
-      extends RuntimeException(s"Error when sending message: ${cause.getMessage}", cause)
+  case class SendMessageError(cause: Throwable, mData: Option[String] = None)
+      extends RuntimeException(
+        s"Error when sending message: ${cause.getMessage};${mData.map(e => s" DATA:'$e'").getOrElse("")}",
+        cause
+      )
       with TransportError
 
   case class ParseResponse(cause: Throwable)